release prep: portable install + packaging hygiene (A3+A5)
install.sh now GENERATES the systemd unit with ExecStart/
WorkingDirectory/PATH derived from the actual checkout and node binary —
the static astrolabe.service (kept as a commented template) hardcoded
the author's ~/claude/zodiac path and a NixOS-specific PATH, so a Linux
install anywhere else pointed systemd at a directory that doesn't
exist. Cargo.toml gains license/repository/authors/readme/keywords/
categories. cargo clippy --all-targets is now clean (24 warnings fixed:
auto-fixables applied, clamp/sort_by_key/while-let rewrites, struct-init
tests, targeted allows for the painter's arg count and a few tuple-typed
fields). Web build artifacts were already gitignored and uncommitted.

+++ src/server.rs
@@ -1203,7 +1203,7 @@ fn host_uptime_ms() -> Option<u64> {
         .duration_since(std::time::UNIX_EPOCH)
         .ok()?
         .as_secs() as i64;
-    Some(((now - tv.tv_sec as i64).max(0) as u64) * 1000)
+    Some(((now - tv.tv_sec).max(0) as u64) * 1000)
 }
 
 /// 1-minute load average over the core count, as a percentage. A
